Added support for the SocialConnection collection of links for each person Created entities for SocialService and SocialConnection Updated DbPerson@hashCode/equals to only use the primary key Don't serialize non-null values

+++ server/src/test/java/org/spring2885/server/api/PersonsApiTest.java
@@ -74,7 +74,7 @@ public class PersonsApiTest {
         otherDbPerson = createDbPerson(21, "other@example.com", "user 21");
     }
     
-    static DbPerson createDbPerson(int id, String email, String aboutMe) {
+    static DbPerson createDbPerson(long id, String email, String aboutMe) {
     	DbPerson p = new DbPerson();
     	p.setId(id);
     	p.setEmail(email);
